[project @ Fixed compilation problems on x86_64]
Original author: Armin Burgmeier <armin@0x539.de>
Date: 2005-03-17 22:01:52+00:00

+++ src/delete_record.cpp
@@ -58,7 +58,7 @@ net6::packet obby::delete_record::to_packet()
 {
 	net6::packet pack("obby_record");
 	pack << "delete" << m_id << m_document << m_revision << m_from
-	     << m_pos << m_text;
+	     << static_cast<unsigned int>(m_pos) << m_text;
 	return pack;
 }
